+++ app/frontend/components/__init__.py
@@ -1,0 +1,14 @@
+from .base import BaseComponent
+from .locale import LocaleStore, _l, _l_dt, _l_dt_relative
+from .primitives import Box, Button, Store
+
+__all__ = [
+    "BaseComponent",
+    "Box",
+    "Button",
+    "LocaleStore",
+    "Store",
+    "_l",
+    "_l_dt",
+    "_l_dt_relative",
+]
