+++ app/frontend/components/auth/__init__.py
@@ -1,0 +1,4 @@
+from .login_form import LoginForm
+from .reg_form import RegForm
+
+__all__ = ["LoginForm", "RegForm"]
