--- app/frontend/components/buttons/__init__.py
@@ -1,1 +1,0 @@
-from .session import TerminateAllOtherSessionBtn, TerminateSessionBtn
