--- app/frontend/components/forms/__init__.py
@@ -1,3 +1,0 @@
-from .login import LoginForm
-from .profile import ProfileForm
-from .registration import RegForm
