--- app/frontend/components/forms/login/__init__.py
@@ -1,1 +1,0 @@
-from .form import LoginForm
