--- app/frontend/components/forms/profile/__init__.py
@@ -1,1 +1,0 @@
-from .form import ProfileForm
