--- app/frontend/components/forms/registration/__init__.py
@@ -1,1 +1,0 @@
-from .form import RegForm
