+++ app/frontend/components/primitives/__init__.py
@@ -1,0 +1,5 @@
+from .box import Box
+from .button import Button
+from .store import Store
+
+__all__ = ["Box", "Button", "Store"]
