+++ app/frontend/components/profile/__init__.py
@@ -1,0 +1,5 @@
+from .permissions_tab import PermissionsTab
+from .profile_form import ProfileForm
+from .session_tab import SessionTab
+
+__all__ = ["PermissionsTab", "ProfileForm", "SessionTab"]
