+++ app/frontend/components/shared/__init__.py
@@ -1,0 +1,19 @@
+from .inputs import (
+    PwdInput,
+    RememberMe,
+    UserEmailInput,
+    UserFirstNameInput,
+    UserLastNameInput,
+    UserSexInput,
+)
+from .language_picker import LanguagePicker
+
+__all__ = [
+    "LanguagePicker",
+    "PwdInput",
+    "RememberMe",
+    "UserEmailInput",
+    "UserFirstNameInput",
+    "UserLastNameInput",
+    "UserSexInput",
+]
