+++ app/frontend/components/shared/inputs/__init__.py
@@ -1,0 +1,17 @@
+from .password import PwdInput
+from .user_attr import (
+    RememberMe,
+    UserEmailInput,
+    UserFirstNameInput,
+    UserLastNameInput,
+    UserSexInput,
+)
+
+__all__ = [
+    "PwdInput",
+    "RememberMe",
+    "UserEmailInput",
+    "UserFirstNameInput",
+    "UserLastNameInput",
+    "UserSexInput",
+]
