+++ app/frontend/pages/auth/registration.py
@@ -2,7 +2,7 @@ import dash
 import dash_mantine_components as dmc
 from flask_login import current_user
 
-from app.frontend.components.forms import RegForm
+from app.frontend.components.auth import RegForm
 from app.frontend.components.locale import _l
 
 dash.register_page(__name__, path="/reg")
