+++ app/frontend/pages/profile.py
@@ -4,10 +4,8 @@ import dash_mantine_components as dmc
 from flask_login import current_user
 
 from app.error import UserNotAuthenticated
-from app.frontend.components.forms import ProfileForm
 from app.frontend.components.locale import _l
-from app.frontend.components.permissions_tab import PermissionsTab
-from app.frontend.components.session import SessionTab
+from app.frontend.components.profile import PermissionsTab, ProfileForm, SessionTab
 
 register_page(
     __name__,
